Add finishing touches to JSON spec draft
This is a draft that might become a final draft. It is always open to
changes which is why I say it is a draft.

Additionally, I have to change the Stitch program to comply more to
this specification, as things like appending to blocks from a different
file are different than the current implementation. However, I will
probably do this after writing a Pandoc AST frontend.